move version check into separate module
