+++ tests/test_qdrant_client.py
@@ -191,54 +191,6 @@ def test_client_init():
     assert client.init_options["metadata"] == {"some-rest-meta": "some-value"}
 
 
-@pytest.mark.parametrize(
-    "test_data",
-    [
-        ("1.9.3.dev0", "2.0.1", False, "Diff between major versions = 1, minor versions differ"),
-        ("1.9.0", "2.9.0", False, "Diff between major versions = 1, minor versions are the same"),
-        (
-            "1.1.0",
-            "1.2.9",
-            True,
-            "Diff between major versions == 0, diff between minor versions == 1 (server > client)",
-        ),
-        (
-            "1.2.7",
-            "1.1.8.dev0",
-            True,
-            "Diff between major versions == 0, diff between minor versions == 1 (client > server)",
-        ),
-        (
-            "1.2.1",
-            "1.2.29",
-            True,
-            "Diff between major versions == 0, diff between minor versions == 0",
-        ),
-        ("1.2.0", "1.2.0", True, "Same versions"),
-        (
-            "1.2.0",
-            "1.4.0",
-            False,
-            "Diff between major versions == 0, diff between minor versions > 1 (server > client)",
-        ),
-        (
-            "1.4.0",
-            "1.2.0",
-            False,
-            "Diff between major versions == 0, diff between minor versions > 1 (client > server)",
-        ),
-        ("1.9.0", "3.7.0", False, "Diff between major versions > 1 (server > client)"),
-        ("3.0.0", "1.0.0", False, "Diff between major versions > 1 (client > server)"),
-    ],
-)
-def test_check_versions(test_data):
-    client = QdrantClient()
-    assert (
-        client._client._check_versions(client_version=test_data[0], server_version=test_data[1])
-        is test_data[2]
-    )
-
-
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 @pytest.mark.parametrize("parallel", [1, 2])
 def test_records_upload(prefer_grpc, parallel):
